Update the SConsOptions and SConsValues classes
To remove the dependency on optparse, we need to update to the argparse
module.  This means we need to derive from the ArgumentParser and
update the parsing methods to return the parsed options.  However, we
also need to carry around the parsed values along with the remaining
arguments.  These we store with the parser.  We also need to update the
Namespace (previously the Values) to work with the defaults SCons
expects to be able to find.