Move target collection to after reading SConscripts
We need to allow all of the SConscripts the time to parse optional
arguments.  This means we cannot build the target list until after
reading all of the SConscripts.  We do need to remove the variables as
we scan them from the list of remaining arguments.  For this, we look
for '=' in the argument unless it is part of a long option.  This does
preclude our ability to specify targets that contain a '=' character.